agent/requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.25.0
